archive/requirements.txt
# Python version: 3.11.13
# Generated from: conda environment hsa_agent
# To recreate: conda create -n hsa_agent python=3.11 && pip install -r requirements.txt

# Core dependencies
python-dotenv>=1.0.0
openai>=2.8.0
google-genai>=1.50.1
google-adk>=1.18.0

# Jupyter/IPython support
ipykernel>=7.0.0

# Search functionality (if needed)
google-search-results>=2.4.0

# Note: The following are dependencies of the packages above and will be installed automatically
# google-api-core, google-auth, google-cloud-*, protobuf, httpx, pydantic, etc.
